+++ tests/testthat/EXS_collectRadii.R
@@ -1,10 +1,5 @@
 ## Run script in aaaMakeData4Vignettes.R to create these objects ... make sure
 ## to run the last line that copies the RData files to the tests folder.
 
-load("testdata/Scale_1_DHO.RData")
-dat1 <- dat
-
-load("testdata/Oto140306_DHO.RData")
-dat2 <- dat
-
-rm(dat)
+dat1 <- readRDS("Scale_1_DHO.rds")
+dat2 <- readRDS("Oto140306_DHO.rds")