+++ vignettes/MeasureRadii/aaaMakeData4Vignettes.R
@@ -10,11 +10,11 @@ digitizeRadii("Scale_1.jpg",id="1",reading="DHO3",edgeIsAnnulus=FALSE)
 digitizeRadii("Scale_2.jpg",id="2",reading="DHO",edgeIsAnnulus=FALSE)
 
 #### Some tests of these functions
-showDigitizedImage("Scale_1_DHO.RData")
-showDigitizedImage(c("Scale_1_DHO.RData","Scale_1_DHO2.RData","Scale_1_DHO3.RData"))
+showDigitizedImage("Scale_1_DHO.rds")
+showDigitizedImage(c("Scale_1_DHO.rds","Scale_1_DHO2.rds","Scale_1_DHO3.rds"))
 showDigitizedImage()  # choose one file and then choose the three
-combineData("Scale_1_DHO.RData")
-combineData(c("Scale_1_DHO.RData","Scale_1_DHO2.RData","Scale_1_DHO3.RData"))
+combineData("Scale_1_DHO.rds")
+combineData(c("Scale_1_DHO.rds","Scale_1_DHO2.rds","Scale_1_DHO3.rds"))
 combineData()  # choose one file and then choose the two
 
 ## Process the otolith
@@ -22,13 +22,13 @@ digitizeRadii("Oto140306.jpg",id="140306",reading="DHO",
               description="Used to demonstrate use of scale-bar.",
               scaleBar=TRUE,scaleBarLength=1,edgeIsAnnulus=TRUE,
               windowSize=12)
-showDigitizedImage("Oto140306_DHO.RData",cex.ann=0.7)
-showDigitizedImage("Oto140306_DHO.RData",annuliLabels=1:6)
-showDigitizedImage("Oto140306_DHO.RData",annuliLabels=c(2,5))
+showDigitizedImage("Oto140306_DHO.rds",cex.ann=0.7)
+showDigitizedImage("Oto140306_DHO.rds",annuliLabels=1:6)
+showDigitizedImage("Oto140306_DHO.rds",annuliLabels=c(2,5))
 
 ## Open the otolith and just get the scaling factor from the scale-bar
 ## then use this to supply the scaling factor rather than use the scale-bar
-## siee ift he results are basically the same as above
+## siee if the results are basically the same as above
 (SF <- findScalingFactor("Oto140306.jpg",knownLength=1,windowSize=12) )
 
 digitizeRadii("Oto140306.jpg",id="140306",reading="DHO2",
@@ -39,4 +39,4 @@ digitizeRadii("Oto140306.jpg",id="140306",reading="DHO2",
 
 #### Copy this to test suite so that an error is thrown if something changed
 ####    Need to do this because everything is interactive
-file.copy(listFiles(".RData"),"../../tests/testthat/testdata/",overwrite=TRUE)
+file.copy(listFiles(".rds"),"../../tests/testthat/",overwrite=TRUE)